+++ src/test/converter2/issues/gh2175.ts
@@ -1,0 +1,5 @@
+/**
+ * `undefined` is special in that it has no declarations, so we can have
+ * an export without any declarations so long as it is undefined.
+ */
+export default undefined;